refactor(h0062): split oversized modules and add E2E coverage
Splits three files that exceeded the CLAUDE.md 800-line guideline,
plus adds five new E2E specs covering the H-0062 paths that did not
yet have end-to-end coverage. Pure cleanup + tests, no behavior change.

File splits (all backward compatible via re-exports):
- backends/lizyml.py (1033 lines) -> backends/lizyml/ package:
  - pickle_compat.py (148) -- preflight + version compatibility check
  - serialization.py    (138) -- TuningResult / SearchDim serializers
  - config_compat.py    (195) -- parse_re_tune + task/objective/metric
                                  validators + strip_internal_keys
  - adapter.py          (613) -- LizyMLAdapter class itself
  - __init__.py          (37) -- re-exports the public surface so
                                  ``from lizystudio.backends.lizyml import``
                                  call sites need no changes
- services/training.py (826 lines) -> training.py (548) +
  training_retune.py (345) for retune-specific helpers. The original
  module re-exports ``run_retune`` / ``start_retune_async`` so the
  api layer's import path is unchanged.
- api/jobs.py (766 lines) -> jobs.py (467) + retune.py (328) for
  retune / resume / lineage endpoints. Both modules share the same
  ``router`` instance so URL paths and ``include_router`` wiring are
  unchanged. ``api/jobs.py`` imports ``api.retune`` at the bottom so
  the ``@router.post(...)`` decorator side-effects run on first load.

Test fallout fixes:
- test_lizyml_checkpoint: monkeypatch target moved from the package
  root to ``backends.lizyml.adapter`` (where cloudpickle is now imported).
- test_training_coverage: monkeypatch target for ``_run_retune_subprocess``
  moved from ``services.training`` to ``services.training_retune`` --
  the retune helper lives there now and the re-export in training.py
  binds a different name in a different module namespace.
- test_retune_api: ``_auto_remaining_trials`` import path updated
  from ``api.jobs`` to ``api.retune``.

E2E additions (frontend/tests/e2e/retune-flow.spec.ts):
- B-1 API: cancel during retune frees the parent lock and marks the
  child cancelled. Verifies the per-parent exclusive lock releases on
  any terminal child state and a follow-up retune is accepted.
- B-2 API: second retune on the same parent returns 409 PARENT_LOCKED
  while the first child is still running.
- B-3 API: corrupted model_meta.json on the parent surfaces as 400
  PICKLE_INCOMPATIBLE on the POST itself, not as a failed background
  child job hours later.
- B-4 UI: Re-tune via the actual Workspace UI shows the Lineage panel
  for the resulting parent / child pair and the click-through wires
  back to the WorkspacePage selection (closes the last gap in
  acceptance criterion #13 from a user-facing perspective).
- B-5 UI: grandchild Re-tune button is enabled on a child job
  (Decision flip 2026-04-14 verified end-to-end).

Quality gates:
- Backend: 945 pytest, mypy clean, ruff clean
- Frontend: 1270 vitest, build success, biome 15 warnings (all pre-existing)
- E2E: spec list verified via ``playwright test --list`` (33 cases,
  all 5 new specs recognised). Full execution deferred to CI -- the
  ``setupTuneJob`` helper depends on ``security.ALLOWED_FILES_ROOT``
  being set to permit the /tmp test fixture, which differs across
  local environments.

Outstanding follow-ups (HISTORY.md H-0062 section) reduced:
- Lineage UI wire-in -- closed via E2E B-4
- File splitting -- closed
- E2E for cancel / lock / pickle incompat / lineage UI / grandchild UI
  -- closed
